requirements.txt
PyYAML>=6.0
Pillow>=9.0
numpy>=1.24
opencv-python>=4.7
tqdm>=4.65
